onion: Move cli onion tool to devtools/onion
